removing enums from the entity.enums package and moving to todo.enums, adding class TaskDTO
